move serial test to main
